// src/websocket-client.d.ts
export type WsKey = string | 'spot' | 'margin' | 'usdmfutures' | 'coinmfutures' | 'options';
export declare interface WebsocketClient {
    on(event: 'reply', listener: (event: WsResponse) => void): this;
    on(event: 'message', listener: (event: WsRawMessage) => void): this;
    on(event: 'formattedMessage', listener: (event: WsFormattedMessage) => void): this;
    on(event: 'formattedUserDataMessage', listener: (event: WsUserDataEvents) => void): this;
    on(event: 'error', listener: (event: {
        wsKey: WsKey;
        error: any;
        rawEvent?: string;
    }) => void): this;
    on(event: 'open' | 'reconnected' | 'reconnecting' | 'close', listener: (event: {
        wsKey: WsKey;
        ws: WebSocket;
        event?: any;
    }) => void): this;
}